port the changes between commit 4c05c2b09e71336d6b11207e6d12e486b4d2705c and bb9d4e0e56bb8f87701142190c1aa2c60a4a6043 from parity/frontier
