+++ client/rpc-core/src/types/transaction.rs
@@ -94,8 +94,8 @@ impl From<TransactionV2> for Transaction {
 				to: None,
 				value: t.value,
 				gas_price: Some(t.gas_price),
-				max_fee_per_gas: Some(t.gas_price),
-				max_priority_fee_per_gas: Some(t.gas_price),
+				max_fee_per_gas: None,
+				max_priority_fee_per_gas: None,
 				gas: t.gas_limit,
 				input: Bytes(t.clone().input),
 				creates: None,
@@ -119,8 +119,8 @@ impl From<TransactionV2> for Transaction {
 				to: None,
 				value: t.value,
 				gas_price: Some(t.gas_price),
-				max_fee_per_gas: Some(t.gas_price),
-				max_priority_fee_per_gas: Some(t.gas_price),
+				max_fee_per_gas: None,
+				max_priority_fee_per_gas: None,
 				gas: t.gas_limit,
 				input: Bytes(t.clone().input),
 				creates: None,
