+++ client/rpc/Cargo.toml
@@ -44,6 +44,13 @@ fc-rpc-core = { version = "1.1.0-dev", path = "../rpc-core" }
 fp-rpc = { version = "3.0.0-dev", path = "../../primitives/rpc" }
 fp-storage = { version = "2.0.0-dev", path = "../../primitives/storage"}
 
+[dev-dependencies]
+tempfile = "3.3.0"
+substrate-test-runtime-client = { branch = "polkadot-v0.9.18", git = "https://github.com/paritytech/substrate" }
+frontier-template-runtime = { path = "../../template/runtime", default-features = false, features = ["std", "aura"] }
+sc-block-builder = { branch = "polkadot-v0.9.18", git = "https://github.com/paritytech/substrate" }
+sp-consensus = { branch = "polkadot-v0.9.18", git = "https://github.com/paritytech/substrate" }
+
 [features]
 default = ["rpc_binary_search_estimate"]
 rpc_binary_search_estimate = []
