+++ client/rpc/src/lib.rs
@@ -20,6 +20,7 @@ mod eth;
 mod eth_pubsub;
 mod net;
 mod overrides;
+mod signer;
 mod tx_pool;
 mod web3;
 
@@ -32,18 +33,15 @@ pub use self::{
 		SchemaV3Override, StorageOverride,
 	},
 	tx_pool::TxPoolApi,
+	signer::{EthDevSigner, EthSigner},
 	web3::Web3Api,
 };
 
 pub use ethereum::TransactionV2 as EthereumTransaction;
-use ethereum_types::{H160, H256};
-use evm::{ExitError, ExitReason};
 pub use fc_rpc_core::{
-	types::TransactionMessage, EthApiServer, EthFilterApiServer, EthPubSubApiServer, NetApiServer,
+	EthApiServer, EthFilterApiServer, EthPubSubApiServer, NetApiServer,
 	TxPoolApiServer, Web3ApiServer,
 };
-use jsonrpc_core::{Error, ErrorCode, Value};
-use sp_core::hashing::keccak_256;
 
 pub mod frontier_backend_client {
 	use super::internal_err;
@@ -191,56 +189,11 @@ pub mod frontier_backend_client {
 	}
 }
 
-pub fn internal_err<T: ToString>(message: T) -> Error {
-	Error {
-		code: ErrorCode::InternalError,
+pub fn internal_err<T: ToString>(message: T) -> jsonrpc_core::Error {
+	jsonrpc_core::Error {
+		code: jsonrpc_core::ErrorCode::InternalError,
 		message: message.to_string(),
 		data: None,
-	}
-}
-
-pub fn error_on_execution_failure(reason: &ExitReason, data: &[u8]) -> Result<(), Error> {
-	match reason {
-		ExitReason::Succeed(_) => Ok(()),
-		ExitReason::Error(e) => {
-			if *e == ExitError::OutOfGas {
-				// `ServerError(0)` will be useful in estimate gas
-				return Err(Error {
-					code: ErrorCode::ServerError(0),
-					message: format!("out of gas"),
-					data: None,
-				});
-			}
-			Err(Error {
-				code: ErrorCode::InternalError,
-				message: format!("evm error: {:?}", e),
-				data: Some(Value::String("0x".to_string())),
-			})
-		}
-		ExitReason::Revert(_) => {
-			let mut message = "VM Exception while processing transaction: revert".to_string();
-			// A minimum size of error function selector (4) + offset (32) + string length (32)
-			// should contain a utf-8 encoded revert reason.
-			if data.len() > 68 {
-				let message_len = data[36..68].iter().sum::<u8>();
-				if data.len() >= 68 + message_len as usize {
-					let body: &[u8] = &data[68..68 + message_len as usize];
-					if let Ok(reason) = std::str::from_utf8(body) {
-						message = format!("{} {}", message, reason.to_string());
-					}
-				}
-			}
-			Err(Error {
-				code: ErrorCode::InternalError,
-				message,
-				data: Some(Value::String(hex::encode(data))),
-			})
-		}
-		ExitReason::Fatal(e) => Err(Error {
-			code: ErrorCode::InternalError,
-			message: format!("evm fatal: {:?}", e),
-			data: Some(Value::String("0x".to_string())),
-		}),
 	}
 }
 
@@ -268,138 +221,4 @@ pub fn public_key(transaction: &EthereumTransaction) -> Result<[u8; 64], sp_io::
 		}
 	}
 	sp_io::crypto::secp256k1_ecdsa_recover(&sig, &msg)
-}
-
-/// A generic Ethereum signer.
-pub trait EthSigner: Send + Sync {
-	/// Available accounts from this signer.
-	fn accounts(&self) -> Vec<H160>;
-	/// Sign a transaction message using the given account in message.
-	fn sign(
-		&self,
-		message: TransactionMessage,
-		address: &H160,
-	) -> Result<EthereumTransaction, Error>;
-}
-
-pub struct EthDevSigner {
-	keys: Vec<libsecp256k1::SecretKey>,
-}
-
-impl EthDevSigner {
-	pub fn new() -> Self {
-		Self {
-			keys: vec![libsecp256k1::SecretKey::parse(&[
-				0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11,
-				0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11, 0x11,
-				0x11, 0x11, 0x11, 0x11,
-			])
-			.expect("Test key is valid; qed")],
-		}
-	}
-}
-
-fn secret_key_address(secret: &libsecp256k1::SecretKey) -> H160 {
-	let public = libsecp256k1::PublicKey::from_secret_key(secret);
-	public_key_address(&public)
-}
-
-fn public_key_address(public: &libsecp256k1::PublicKey) -> H160 {
-	let mut res = [0u8; 64];
-	res.copy_from_slice(&public.serialize()[1..65]);
-	H160::from(H256::from(keccak_256(&res)))
-}
-
-impl EthSigner for EthDevSigner {
-	fn accounts(&self) -> Vec<H160> {
-		self.keys.iter().map(secret_key_address).collect()
-	}
-
-	fn sign(
-		&self,
-		message: TransactionMessage,
-		address: &H160,
-	) -> Result<EthereumTransaction, Error> {
-		let mut transaction = None;
-
-		for secret in &self.keys {
-			let key_address = secret_key_address(secret);
-
-			if &key_address == address {
-				match message {
-					TransactionMessage::Legacy(m) => {
-						let signing_message = libsecp256k1::Message::parse_slice(&m.hash()[..])
-							.map_err(|_| internal_err("invalid signing message"))?;
-						let (signature, recid) = libsecp256k1::sign(&signing_message, secret);
-						let v = match m.chain_id {
-							None => 27 + recid.serialize() as u64,
-							Some(chain_id) => 2 * chain_id + 35 + recid.serialize() as u64,
-						};
-						let rs = signature.serialize();
-						let r = H256::from_slice(&rs[0..32]);
-						let s = H256::from_slice(&rs[32..64]);
-						transaction =
-							Some(EthereumTransaction::Legacy(ethereum::LegacyTransaction {
-								nonce: m.nonce,
-								gas_price: m.gas_price,
-								gas_limit: m.gas_limit,
-								action: m.action,
-								value: m.value,
-								input: m.input.clone(),
-								signature: ethereum::TransactionSignature::new(v, r, s)
-									.ok_or(internal_err("signer generated invalid signature"))?,
-							}));
-					}
-					TransactionMessage::EIP2930(m) => {
-						let signing_message = libsecp256k1::Message::parse_slice(&m.hash()[..])
-							.map_err(|_| internal_err("invalid signing message"))?;
-						let (signature, recid) = libsecp256k1::sign(&signing_message, secret);
-						let rs = signature.serialize();
-						let r = H256::from_slice(&rs[0..32]);
-						let s = H256::from_slice(&rs[32..64]);
-						transaction =
-							Some(EthereumTransaction::EIP2930(ethereum::EIP2930Transaction {
-								chain_id: m.chain_id,
-								nonce: m.nonce,
-								gas_price: m.gas_price,
-								gas_limit: m.gas_limit,
-								action: m.action,
-								value: m.value,
-								input: m.input.clone(),
-								access_list: m.access_list,
-								odd_y_parity: recid.serialize() != 0,
-								r,
-								s,
-							}));
-					}
-					TransactionMessage::EIP1559(m) => {
-						let signing_message = libsecp256k1::Message::parse_slice(&m.hash()[..])
-							.map_err(|_| internal_err("invalid signing message"))?;
-						let (signature, recid) = libsecp256k1::sign(&signing_message, secret);
-						let rs = signature.serialize();
-						let r = H256::from_slice(&rs[0..32]);
-						let s = H256::from_slice(&rs[32..64]);
-						transaction =
-							Some(EthereumTransaction::EIP1559(ethereum::EIP1559Transaction {
-								chain_id: m.chain_id,
-								nonce: m.nonce,
-								max_priority_fee_per_gas: m.max_priority_fee_per_gas,
-								max_fee_per_gas: m.max_fee_per_gas,
-								gas_limit: m.gas_limit,
-								action: m.action,
-								value: m.value,
-								input: m.input.clone(),
-								access_list: m.access_list,
-								odd_y_parity: recid.serialize() != 0,
-								r,
-								s,
-							}));
-					}
-				}
-				break;
-			}
-		}
-
-		transaction.ok_or(internal_err("signer not available"))
-	}
 }
