+++ client/rpc/src/overrides/schema_v2_override.rs
@@ -16,22 +16,24 @@
 // You should have received a copy of the GNU General Public License
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
+use std::{marker::PhantomData, sync::Arc};
+
 use codec::Decode;
 use ethereum_types::{H160, H256, U256};
-use fp_rpc::TransactionStatus;
-use sc_client_api::backend::{AuxStore, Backend, StateBackend, StorageProvider};
+
+use sc_client_api::backend::{Backend, StateBackend, StorageProvider};
 use sp_api::BlockId;
-use sp_blockchain::{Error as BlockChainError, HeaderBackend, HeaderMetadata};
 use sp_runtime::{
 	traits::{BlakeTwo256, Block as BlockT},
 	Permill,
 };
 use sp_storage::StorageKey;
-use std::{marker::PhantomData, sync::Arc};
+
+use fp_rpc::TransactionStatus;
 
 use super::{blake2_128_extend, storage_prefix_build, StorageOverride};
 
-/// An override for runtimes that use Schema V1
+/// An override for runtimes that use Schema V2
 pub struct SchemaV2Override<B: BlockT, C, BE> {
 	client: Arc<C>,
 	_marker: PhantomData<(B, BE)>,
@@ -48,12 +50,10 @@ impl<B: BlockT, C, BE> SchemaV2Override<B, C, BE> {
 
 impl<B, C, BE> SchemaV2Override<B, C, BE>
 where
-	C: StorageProvider<B, BE> + AuxStore,
-	C: HeaderBackend<B> + HeaderMetadata<B, Error = BlockChainError> + 'static,
+	B: BlockT<Hash = H256> + Send + Sync + 'static,
+	C: StorageProvider<B, BE> + Send + Sync + 'static,
 	BE: Backend<B> + 'static,
 	BE::State: StateBackend<BlakeTwo256>,
-	B: BlockT<Hash = H256> + Send + Sync + 'static,
-	C: Send + Sync + 'static,
 {
 	fn query_storage<T: Decode>(&self, id: &BlockId<B>, key: &StorageKey) -> Option<T> {
 		if let Ok(Some(data)) = self.client.storage(id, key) {
@@ -65,26 +65,22 @@ where
 	}
 }
 
-impl<Block, C, BE> StorageOverride<Block> for SchemaV2Override<Block, C, BE>
+impl<B, C, BE> StorageOverride<B> for SchemaV2Override<B, C, BE>
 where
-	C: StorageProvider<Block, BE>,
-	C: AuxStore,
-	C: HeaderBackend<Block>,
-	C: HeaderMetadata<Block, Error = BlockChainError> + 'static,
-	BE: Backend<Block> + 'static,
+	B: BlockT<Hash = H256> + Send + Sync + 'static,
+	C: StorageProvider<B, BE> + Send + Sync + 'static,
+	BE: Backend<B> + 'static,
 	BE::State: StateBackend<BlakeTwo256>,
-	Block: BlockT<Hash = H256> + Send + Sync + 'static,
-	C: Send + Sync + 'static,
 {
 	/// For a given account address, returns pallet_evm::AccountCodes.
-	fn account_code_at(&self, block: &BlockId<Block>, address: H160) -> Option<Vec<u8>> {
+	fn account_code_at(&self, block: &BlockId<B>, address: H160) -> Option<Vec<u8>> {
 		let mut key: Vec<u8> = storage_prefix_build(b"EVM", b"AccountCodes");
 		key.extend(blake2_128_extend(address.as_bytes()));
 		self.query_storage::<Vec<u8>>(block, &StorageKey(key))
 	}
 
 	/// For a given account address and index, returns pallet_evm::AccountStorages.
-	fn storage_at(&self, block: &BlockId<Block>, address: H160, index: U256) -> Option<H256> {
+	fn storage_at(&self, block: &BlockId<B>, address: H160, index: U256) -> Option<H256> {
 		let tmp: &mut [u8; 32] = &mut [0; 32];
 		index.to_big_endian(tmp);
 
@@ -96,7 +92,7 @@ where
 	}
 
 	/// Return the current block.
-	fn current_block(&self, block: &BlockId<Block>) -> Option<ethereum::BlockV2> {
+	fn current_block(&self, block: &BlockId<B>) -> Option<ethereum::BlockV2> {
 		self.query_storage::<ethereum::BlockV2>(
 			block,
 			&StorageKey(storage_prefix_build(b"Ethereum", b"CurrentBlock")),
@@ -104,7 +100,7 @@ where
 	}
 
 	/// Return the current receipt.
-	fn current_receipts(&self, block: &BlockId<Block>) -> Option<Vec<ethereum::ReceiptV3>> {
+	fn current_receipts(&self, block: &BlockId<B>) -> Option<Vec<ethereum::ReceiptV3>> {
 		self.query_storage::<Vec<ethereum::ReceiptV0>>(
 			block,
 			&StorageKey(storage_prefix_build(b"Ethereum", b"CurrentReceipts")),
@@ -125,10 +121,7 @@ where
 	}
 
 	/// Return the current transaction status.
-	fn current_transaction_statuses(
-		&self,
-		block: &BlockId<Block>,
-	) -> Option<Vec<TransactionStatus>> {
+	fn current_transaction_statuses(&self, block: &BlockId<B>) -> Option<Vec<TransactionStatus>> {
 		self.query_storage::<Vec<TransactionStatus>>(
 			block,
 			&StorageKey(storage_prefix_build(
@@ -139,15 +132,15 @@ where
 	}
 
 	/// Return the base fee at the given height.
-	fn base_fee(&self, block: &BlockId<Block>) -> Option<U256> {
+	fn base_fee(&self, block: &BlockId<B>) -> Option<U256> {
 		self.query_storage::<U256>(
 			block,
 			&StorageKey(storage_prefix_build(b"BaseFee", b"BaseFeePerGas")),
 		)
 	}
 
-	/// Prior to eip-1559 there is no base fee.
-	fn elasticity(&self, block: &BlockId<Block>) -> Option<Permill> {
+	/// Return the elasticity at the given height.
+	fn elasticity(&self, block: &BlockId<B>) -> Option<Permill> {
 		let default_elasticity = Some(Permill::from_parts(125_000));
 		let elasticity = self.query_storage::<Permill>(
 			block,
@@ -160,7 +153,7 @@ where
 		}
 	}
 
-	fn is_eip1559(&self, _block: &BlockId<Block>) -> bool {
+	fn is_eip1559(&self, _block: &BlockId<B>) -> bool {
 		true
 	}
 }
