+++ frame/evm/src/benchmarking.rs
@@ -76,6 +76,7 @@ benchmarks! {
 
 		let value = U256::default();
 		let gas_limit_create: u64 = 1_250_000 * 1_000_000_000;
+		let is_transactional = true;
 		let create_runner_results = T::Runner::create(
 			caller,
 			contract_bytecode,
@@ -85,6 +86,7 @@ benchmarks! {
 			None,
 			Some(nonce_as_u256),
 			Vec::new(),
+			is_transactional,
 			T::config(),
 		);
 		assert_eq!(create_runner_results.is_ok(), true, "create() failed");
@@ -106,6 +108,7 @@ benchmarks! {
 		nonce = nonce + 1;
 		let nonce_as_u256: U256 = nonce.into();
 
+		let is_transactional = true;
 		let call_runner_results = T::Runner::call(
 			caller,
 			contract_address,
@@ -116,6 +119,7 @@ benchmarks! {
 			None,
 			Some(nonce_as_u256),
 			Vec::new(),
+			is_transactional,
 			T::config(),
 		);
 		assert_eq!(call_runner_results.is_ok(), true, "call() failed");
