+++ frame/evm/src/runner/mod.rs
@@ -35,6 +35,7 @@ pub trait Runner<T: Config> {
 		max_priority_fee_per_gas: Option<U256>,
 		nonce: Option<U256>,
 		access_list: Vec<(H160, Vec<H256>)>,
+		is_transactional: bool,
 		config: &evm::Config,
 	) -> Result<CallInfo, Self::Error>;
 
@@ -47,6 +48,7 @@ pub trait Runner<T: Config> {
 		max_priority_fee_per_gas: Option<U256>,
 		nonce: Option<U256>,
 		access_list: Vec<(H160, Vec<H256>)>,
+		is_transactional: bool,
 		config: &evm::Config,
 	) -> Result<CreateInfo, Self::Error>;
 
@@ -60,6 +62,7 @@ pub trait Runner<T: Config> {
 		max_priority_fee_per_gas: Option<U256>,
 		nonce: Option<U256>,
 		access_list: Vec<(H160, Vec<H256>)>,
+		is_transactional: bool,
 		config: &evm::Config,
 	) -> Result<CreateInfo, Self::Error>;
 }
