+++ frame/evm/src/runner/stack.rs
@@ -52,6 +52,7 @@ impl<T: Config> Runner<T> {
 		nonce: Option<U256>,
 		config: &'config evm::Config,
 		precompiles: &'precompiles T::PrecompilesType,
+		is_transactional: bool,
 		f: F,
 	) -> Result<ExecutionInfo<R>, Error<T>>
 	where
@@ -65,17 +66,54 @@ impl<T: Config> Runner<T> {
 		) -> (ExitReason, R),
 	{
 		let base_fee = T::FeeCalculator::min_gas_price();
-		// Gas price check is skipped when performing a gas estimation.
-		let max_fee_per_gas = match max_fee_per_gas {
-			Some(max_fee_per_gas) => {
+		let max_fee_per_gas = match (max_fee_per_gas, is_transactional) {
+			(Some(max_fee_per_gas), _) => {
 				ensure!(max_fee_per_gas >= base_fee, Error::<T>::GasPriceTooLow);
 				max_fee_per_gas
 			}
-			None => Default::default(),
+			// Gas price check is skipped for non-transactional calls that don't
+			// define a `max_fee_per_gas` input.
+			(None, false) => Default::default(),
+			_ => return Err(Error::<T>::GasPriceTooLow),
 		};
 
+		if let Some(max_priority_fee) = max_priority_fee_per_gas {
+			ensure!(
+				max_fee_per_gas >= max_priority_fee,
+				Error::<T>::GasPriceTooLow
+			);
+		}
+
+		// After eip-1559 we make sure the account can pay both the evm execution and priority fees.
+		let total_fee = max_fee_per_gas
+			.checked_mul(U256::from(gas_limit))
+			.ok_or(Error::<T>::FeeOverflow)?;
+
+		let total_payment = value
+			.checked_add(total_fee)
+			.ok_or(Error::<T>::PaymentOverflow)?;
+		let source_account = Pallet::<T>::account_basic(&source);
+		// Account balance check is skipped if fee is Zero.
+		// This case is previously verified to only happen on either:
+		// 	- Non-transactional calls.
+		//	- BaseFee is configured to be Zero.
+		if total_fee > U256::zero() {
+			ensure!(
+				source_account.balance >= total_payment,
+				Error::<T>::BalanceLow
+			);
+		}
+
+		if let Some(nonce) = nonce {
+			ensure!(source_account.nonce == nonce, Error::<T>::InvalidNonce);
+		}
+
+		// Deduct fee from the `source` account. Returns `None` if `total_fee` is Zero.
+		let fee = T::OnChargeTransaction::withdraw_fee(&source, total_fee)?;
+
+		// Execute the EVM call.
 		let vicinity = Vicinity {
-			gas_price: max_fee_per_gas,
+			gas_price: base_fee,
 			origin: source,
 		};
 
@@ -83,44 +121,15 @@ impl<T: Config> Runner<T> {
 		let state = SubstrateStackState::new(&vicinity, metadata);
 		let mut executor = StackExecutor::new_with_precompiles(state, config, precompiles);
 
-		// After eip-1559 we make sure the account can pay both the evm execution and priority fees.
-		let max_base_fee = max_fee_per_gas
-			.checked_mul(U256::from(gas_limit))
-			.ok_or(Error::<T>::FeeOverflow)?;
-		let max_priority_fee = if let Some(max_priority_fee) = max_priority_fee_per_gas {
-			max_priority_fee
-				.checked_mul(U256::from(gas_limit))
-				.ok_or(Error::<T>::FeeOverflow)?
-		} else {
-			U256::zero()
-		};
-
-		let total_fee = max_base_fee
-			.checked_add(max_priority_fee)
-			.ok_or(Error::<T>::FeeOverflow)?;
-
-		let total_payment = value
-			.checked_add(total_fee)
-			.ok_or(Error::<T>::PaymentOverflow)?;
-		let source_account = Pallet::<T>::account_basic(&source);
-		ensure!(
-			source_account.balance >= total_payment,
-			Error::<T>::BalanceLow
-		);
-
-		if let Some(nonce) = nonce {
-			ensure!(source_account.nonce == nonce, Error::<T>::InvalidNonce);
-		}
-		// Deduct fee from the `source` account.
-		let fee = T::OnChargeTransaction::withdraw_fee(&source, total_fee)?;
-
-		// Execute the EVM call.
 		let (reason, retv) = f(&mut executor);
 
+		// Post execution.
 		let used_gas = U256::from(executor.used_gas());
 		let (actual_fee, actual_priority_fee) =
 			if let Some(max_priority_fee) = max_priority_fee_per_gas {
-				let actual_priority_fee = max_priority_fee
+				let actual_priority_fee = max_fee_per_gas
+					.saturating_sub(base_fee)
+					.min(max_priority_fee)
 					.checked_mul(U256::from(used_gas))
 					.ok_or(Error::<T>::FeeOverflow)?;
 				let actual_fee = executor
@@ -133,12 +142,13 @@ impl<T: Config> Runner<T> {
 			};
 		log::debug!(
 			target: "evm",
-			"Execution {:?} [source: {:?}, value: {}, gas_limit: {}, actual_fee: {}]",
+			"Execution {:?} [source: {:?}, value: {}, gas_limit: {}, actual_fee: {}, is_transactional: {}]",
 			reason,
 			source,
 			value,
 			gas_limit,
-			actual_fee
+			actual_fee,
+			is_transactional
 		);
 		// The difference between initially withdrawn and the actual cost is refunded.
 		//
@@ -156,11 +166,11 @@ impl<T: Config> Runner<T> {
 		// |        5 |        2 |
 		// +----------+----------+
 		//
-		// Initially withdrawn (10 + 6) * 20 = 320.
+		// Initially withdrawn 10 * 20 = 200.
 		// Actual cost (2 + 6) * 5 = 40.
-		// Refunded 320 - 40 = 280.
+		// Refunded 200 - 40 = 160.
 		// Tip 5 * 6 = 30.
-		// Burned 320 - (280 + 30) = 10. Which is equivalent to gas_used * base_fee.
+		// Burned 200 - (160 + 30) = 10. Which is equivalent to gas_used * base_fee.
 		if let Some(actual_priority_fee) = actual_priority_fee {
 			T::OnChargeTransaction::correct_and_deposit_fee(
 				&source,
@@ -228,6 +238,7 @@ impl<T: Config> RunnerT<T> for Runner<T> {
 		max_priority_fee_per_gas: Option<U256>,
 		nonce: Option<U256>,
 		access_list: Vec<(H160, Vec<H256>)>,
+		is_transactional: bool,
 		config: &evm::Config,
 	) -> Result<CallInfo, Self::Error> {
 		let precompiles = T::PrecompilesValue::get();
@@ -240,6 +251,7 @@ impl<T: Config> RunnerT<T> for Runner<T> {
 			nonce,
 			config,
 			&precompiles,
+			is_transactional,
 			|executor| executor.transact_call(source, target, value, input, gas_limit, access_list),
 		)
 	}
@@ -253,6 +265,7 @@ impl<T: Config> RunnerT<T> for Runner<T> {
 		max_priority_fee_per_gas: Option<U256>,
 		nonce: Option<U256>,
 		access_list: Vec<(H160, Vec<H256>)>,
+		is_transactional: bool,
 		config: &evm::Config,
 	) -> Result<CreateInfo, Self::Error> {
 		let precompiles = T::PrecompilesValue::get();
@@ -265,6 +278,7 @@ impl<T: Config> RunnerT<T> for Runner<T> {
 			nonce,
 			config,
 			&precompiles,
+			is_transactional,
 			|executor| {
 				let address = executor.create_address(evm::CreateScheme::Legacy { caller: source });
 				let (reason, _) =
@@ -284,6 +298,7 @@ impl<T: Config> RunnerT<T> for Runner<T> {
 		max_priority_fee_per_gas: Option<U256>,
 		nonce: Option<U256>,
 		access_list: Vec<(H160, Vec<H256>)>,
+		is_transactional: bool,
 		config: &evm::Config,
 	) -> Result<CreateInfo, Self::Error> {
 		let precompiles = T::PrecompilesValue::get();
@@ -297,6 +312,7 @@ impl<T: Config> RunnerT<T> for Runner<T> {
 			nonce,
 			config,
 			&precompiles,
+			is_transactional,
 			|executor| {
 				let address = executor.create_address(evm::CreateScheme::Create2 {
 					caller: source,
