+++ frame/evm/src/tests.rs
@@ -48,6 +48,11 @@ pub fn new_test_ext() -> sp_io::TestExternalities {
 		H160::from_str("1000000000000000000000000000000000000003").unwrap(),
 		AccountId32::new([4u8; 32]),
 	);
+	account_pairs.insert(
+		H160::from_str("1000000000000000000000000000000000000004").unwrap(),
+		AccountId32::new([5u8; 32]),
+	);
+
 
 	let mut accounts = BTreeMap::new();
 	accounts.insert(
@@ -232,7 +237,7 @@ fn author_should_get_tip() {
 			Vec::new(),
 			U256::from(1),
 			1000000,
-			U256::from(1_000_000_000),
+			U256::from(2_000_000_000),
 			Some(U256::from(1)),
 			None,
 			Vec::new(),
@@ -303,29 +308,81 @@ fn refunds_and_priority_should_work() {
 		let evm_addr = H160::from_str("1000000000000000000000000000000000000001").unwrap();
 		let account_id = <Test as Config>::AddressMapping::into_account_id(evm_addr);
 		let before_call = EVM::account_basic(&evm_addr).balance;
-		let tip = 5;
-		// The tip is deducted but never refunded to the caller.
+		// We deliberately set a base fee + max tip > max fee.
+		// The effective priority tip will be 1GWEI instead 1.5GWEI:
+		// 		(max_fee_per_gas - base_fee).min(max_priority_fee)
+		//		(2 - 1).min(1.5)
+		let tip = U256::from(1_500_000_000);
+		let max_fee_per_gas = U256::from(2_000_000_000);
+		let used_gas = U256::from(21_000);
 		assert_ok!(EVM::call(
 			Origin::signed(account_id),
 			evm_addr,
 			H160::from_str("1000000000000000000000000000000000000003").unwrap(),
 			Vec::new(),
-			U256::from_str("0xfffffffffffff").unwrap(),
+			U256::from(1),
 			1000000,
-			U256::from(2_000_000_000),
-			Some(U256::from(tip)),
+			max_fee_per_gas,
+			Some(tip),
 			None,
 			Vec::new(),
 		));
-		let tip = tip * 21000;
-		let total_cost = (U256::from(21_000) * <Test as Config>::FeeCalculator::min_gas_price())
-			+ U256::from_str("0xfffffffffffff").unwrap()
-			+ U256::from(tip);
+		let base_fee = <Test as Config>::FeeCalculator::min_gas_price();
+		let actual_tip = (max_fee_per_gas - base_fee).min(tip) * used_gas;
+		let total_cost = (used_gas * base_fee) + U256::from(actual_tip) + U256::from(1);
 		let after_call = EVM::account_basic(&evm_addr).balance;
+		// The tip is deducted but never refunded to the caller.
 		assert_eq!(after_call, before_call - total_cost);
 
 		let after_tip = EVM::account_basic(&author).balance;
-		assert_eq!(after_tip, (before_tip + tip));
+		assert_eq!(after_tip, (before_tip + actual_tip.low_u128()));
+	});
+}
+
+#[test]
+fn call_should_fail_with_priority_greater_than_max_fee() {
+	new_test_ext().execute_with(|| {
+		// Max priority greater than max fee should fail.
+		let tip: u128 = 1_100_000_000;
+		let evm_addr = H160::from_str("1000000000000000000000000000000000000002").unwrap();
+		let account_id = <Test as Config>::AddressMapping::into_account_id(evm_addr);
+		let result = EVM::call(
+			Origin::signed(account_id),
+			evm_addr,
+			H160::from_str("1000000000000000000000000000000000000003").unwrap(),
+			Vec::new(),
+			U256::from(1),
+			1000000,
+			U256::from(1_000_000_000),
+			Some(U256::from(tip)),
+			None,
+			Vec::new(),
+		);
+		assert!(result.is_err());
+	});
+}
+
+#[test]
+fn call_should_succeed_with_priority_equal_to_max_fee() {
+	new_test_ext().execute_with(|| {
+		let tip: u128 = 1_000_000_000;
+		let evm_addr = H160::from_str("1000000000000000000000000000000000000002").unwrap();
+		let account_id = <Test as Config>::AddressMapping::into_account_id(evm_addr);
+		// Mimics the input for pre-eip-1559 transaction types where `gas_price`
+		// is used for both `max_fee_per_gas` and `max_priority_fee_per_gas`.
+		let result = EVM::call(
+			Origin::signed(account_id),
+			evm_addr,
+			H160::from_str("1000000000000000000000000000000000000003").unwrap(),
+			Vec::new(),
+			U256::from(1),
+			1000000,
+			U256::from(1_000_000_000),
+			Some(U256::from(tip)),
+			None,
+			Vec::new(),
+		);
+		assert!(result.is_ok());
 	});
 }
 
@@ -352,5 +409,123 @@ fn handle_sufficient_reference() {
 		let account_2 = frame_system::Account::<Test>::get(substrate_addr_2);
 		// We decreased the sufficient reference by 1 on removing the account.
 		assert_eq!(account_2.sufficients, 0);
+	});
+}
+
+#[test]
+fn runner_non_transactional_calls_with_non_balance_accounts_is_ok_without_gas_price() {
+	// Expect to skip checks for gas price and account balance when both:
+	//	- The call is non transactional (`is_transactional == false`).
+	//	- The `max_fee_per_gas` is None.
+	new_test_ext().execute_with(|| {
+		let non_balance_account =
+			H160::from_str("7700000000000000000000000000000000000001").unwrap();
+		assert_eq!(
+			EVM::account_basic(&non_balance_account).balance,
+			U256::zero()
+		);
+		let _ = <Test as Config>::Runner::call(
+			non_balance_account,
+			H160::from_str("1000000000000000000000000000000000000001").unwrap(),
+			Vec::new(),
+			U256::from(1u32),
+			1000000,
+			None,
+			None,
+			None,
+			Vec::new(),
+			false,
+			&<Test as Config>::config().clone(),
+		)
+		.expect("Non transactional call succeeds");
+		assert_eq!(
+			EVM::account_basic(&non_balance_account).balance,
+			U256::zero()
+		);
+	});
+}
+
+#[test]
+fn runner_non_transactional_calls_with_non_balance_accounts_is_err_with_gas_price() {
+	// In non transactional calls where `Some(gas_price)` is defined, expect it to be
+	// checked against the `BaseFee`, and expect the account to have enough balance
+	// to pay for the call.
+	new_test_ext().execute_with(|| {
+		let non_balance_account =
+			H160::from_str("7700000000000000000000000000000000000001").unwrap();
+		assert_eq!(
+			EVM::account_basic(&non_balance_account).balance,
+			U256::zero()
+		);
+		let res = <Test as Config>::Runner::call(
+			non_balance_account,
+			H160::from_str("1000000000000000000000000000000000000001").unwrap(),
+			Vec::new(),
+			U256::from(1u32),
+			1000000,
+			Some(U256::from(1_000_000_000)),
+			None,
+			None,
+			Vec::new(),
+			false,
+			&<Test as Config>::config().clone(),
+		);
+		assert!(res.is_err());
+	});
+}
+
+#[test]
+fn runner_transactional_call_with_zero_gas_price_fails() {
+	// Transactional calls are rejected when `max_fee_per_gas == None`.
+	new_test_ext().execute_with(|| {
+		let res = <Test as Config>::Runner::call(
+			H160::default(),
+			H160::from_str("1000000000000000000000000000000000000001").unwrap(),
+			Vec::new(),
+			U256::from(1u32),
+			1000000,
+			None,
+			None,
+			None,
+			Vec::new(),
+			true,
+			&<Test as Config>::config().clone(),
+		);
+		assert!(res.is_err());
+	});
+}
+
+#[test]
+fn runner_max_fee_per_gas_gte_max_priority_fee_per_gas() {
+	// Transactional and non transactional calls enforce `max_fee_per_gas >= max_priority_fee_per_gas`.
+	new_test_ext().execute_with(|| {
+		let res = <Test as Config>::Runner::call(
+			H160::default(),
+			H160::from_str("1000000000000000000000000000000000000001").unwrap(),
+			Vec::new(),
+			U256::from(1u32),
+			1000000,
+			Some(U256::from(1_000_000_000)),
+			Some(U256::from(2_000_000_000)),
+			None,
+			Vec::new(),
+			true,
+			&<Test as Config>::config().clone(),
+		);
+		assert!(res.is_err());
+		let res = <Test as Config>::Runner::call(
+			H160::default(),
+			H160::from_str("1000000000000000000000000000000000000001").unwrap(),
+			Vec::new(),
+			U256::from(1u32),
+			1000000,
+			Some(U256::from(1_000_000_000)),
+			Some(U256::from(2_000_000_000)),
+			None,
+			Vec::new(),
+			false,
+			&<Test as Config>::config().clone(),
+		);
+		assert!(res.is_err());
 	});
 }
