+++ rustfmt.toml
@@ -1,2 +1,3 @@
 hard_tabs = true
 imports_granularity = "Crate"
+use_field_init_shorthand = true
