+++ ts-tests/tests/test-pending-pool.ts
@@ -49,3 +49,73 @@ describeWithFrontier("Frontier RPC (Pending Pool)", (context) => {
 		});
 	});
 });
+
+describeWithFrontier("Frontier RPC (Pending Transaction Count)", (context) => {
+	const GENESIS_ACCOUNT = "0x6be02d1d3665660d22ff9624b7be0551ee1ac91b";
+	const GENESIS_ACCOUNT_PRIVATE_KEY = "0x99B3C12287537E38C90A9219D4CB074A89A16E9CDB20BF85728EBD97C343E342";
+	const TEST_ACCOUNT = "0x1111111111111111111111111111111111111111";
+
+	it("should return pending transaction count", async function () {
+		this.timeout(15000);
+
+		// nonce should be 0
+		expect(await context.web3.eth.getTransactionCount(GENESIS_ACCOUNT, 'latest')).to.eq(0);
+
+		var nonce = 0;
+		let sendTransaction = async () => {
+			const tx = await context.web3.eth.accounts.signTransaction(
+				{
+					from: GENESIS_ACCOUNT,
+					to: TEST_ACCOUNT,
+					value: "0x200", // Must be higher than ExistentialDeposit (500)
+					gasPrice: "0x3B9ACA00",
+					gas: "0x100000",
+					nonce: nonce,
+				},
+				GENESIS_ACCOUNT_PRIVATE_KEY
+			);
+			nonce = nonce + 1;
+			return (await customRequest(context.web3, "eth_sendRawTransaction", [tx.rawTransaction])).result
+		};
+
+		{
+			const pending_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", ['pending'])).result;
+			expect(pending_transaction_count).to.eq('0x0');
+		}
+
+		// block 1 should have 1 transaction
+		await sendTransaction();
+		{
+			const pending_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", ['pending'])).result;
+			expect(pending_transaction_count).to.eq('0x1');
+		}
+
+		await createAndFinalizeBlock(context.web3);
+
+		{
+			const pending_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", ['pending'])).result;
+			expect(pending_transaction_count).to.eq('0x0');
+			const processed_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", [1])).result;
+			expect(processed_transaction_count).to.eq('0x1');
+		}
+
+		// block 2 should have 5 transactions
+		for (var _ of Array(5)) {
+			await sendTransaction();
+		}
+
+		{
+			const pending_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", ['pending'])).result;
+			expect(pending_transaction_count).to.eq('0x5');
+		}
+
+		await createAndFinalizeBlock(context.web3);
+
+		{
+			const pending_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", ['pending'])).result;
+			expect(pending_transaction_count).to.eq('0x0');
+			const processed_transaction_count = (await customRequest(context.web3, "eth_getBlockTransactionCountByNumber", [2])).result;
+			expect(processed_transaction_count).to.eq('0x5');
+		}
+	});
+});
