+++ ts-tests/tests/test-transaction-version.ts
@@ -21,14 +21,13 @@ describeWithFrontier("Frontier RPC (Transaction Version)", (context) => {
 		return tx;
 	}
 	
-	step("should handle EIP-2930 transaction type 1", async function () {
+	step("should handle Legacy transaction type 0", async function () {
 		let tx = {
 			from: GENESIS_ACCOUNT,
 			data: TEST_CONTRACT_BYTECODE,
 			value: "0x00",
 			gasPrice: "0x3B9ACA00",
-			type: 1,
-			accessList: [],
+			type: 0,
 			nonce: 0,
 			gasLimit: "0x100000",
 			chainId: 42,
@@ -41,16 +40,22 @@ describeWithFrontier("Frontier RPC (Transaction Version)", (context) => {
 
 		let receipt = await context.web3.eth.getTransactionReceipt(tx_hash);
 		expect(receipt.transactionHash).to.be.eq(tx_hash);
+
+		let transaction_data = await context.web3.eth.getTransaction(tx_hash);
+		expect(transaction_data).to.have.own.property('type');
+		expect(transaction_data).to.not.have.own.property('maxFeePerGas');
+		expect(transaction_data).to.not.have.own.property('maxPriorityFeePerGas');
+
 	});
 	
-	step("should handle EIP-1559 transaction type 2", async function () {
+	
+	step("should handle EIP-2930 transaction type 1", async function () {
 		let tx = {
 			from: GENESIS_ACCOUNT,
 			data: TEST_CONTRACT_BYTECODE,
 			value: "0x00",
-			maxFeePerGas: "0x3B9ACA00",
-			maxPriorityFeePerGas: "0x01",
-			type: 2,
+			gasPrice: "0x3B9ACA00",
+			type: 1,
 			accessList: [],
 			nonce: 1,
 			gasLimit: "0x100000",
@@ -64,5 +69,38 @@ describeWithFrontier("Frontier RPC (Transaction Version)", (context) => {
 
 		let receipt = await context.web3.eth.getTransactionReceipt(tx_hash);
 		expect(receipt.transactionHash).to.be.eq(tx_hash);
+
+		let transaction_data = await context.web3.eth.getTransaction(tx_hash);
+		expect(transaction_data).to.have.own.property('type');
+		expect(transaction_data).to.not.have.own.property('maxFeePerGas');
+		expect(transaction_data).to.not.have.own.property('maxPriorityFeePerGas');
+	});
+	
+	step("should handle EIP-1559 transaction type 2", async function () {
+		let tx = {
+			from: GENESIS_ACCOUNT,
+			data: TEST_CONTRACT_BYTECODE,
+			value: "0x00",
+			maxFeePerGas: "0x3B9ACA00",
+			maxPriorityFeePerGas: "0x01",
+			type: 2,
+			accessList: [],
+			nonce: 2,
+			gasLimit: "0x100000",
+			chainId: 42,
+		};
+		const tx_hash = (await sendTransaction(context, tx)).hash;
+		await createAndFinalizeBlock(context.web3);
+		const latest = await context.web3.eth.getBlock("latest");
+		expect(latest.transactions.length).to.be.eq(1);
+		expect(latest.transactions[0]).to.be.eq(tx_hash);
+
+		let receipt = await context.web3.eth.getTransactionReceipt(tx_hash);
+		expect(receipt.transactionHash).to.be.eq(tx_hash);
+
+		let transaction_data = await context.web3.eth.getTransaction(tx_hash);
+		expect(transaction_data).to.have.own.property('type');
+		expect(transaction_data).to.have.own.property('maxFeePerGas');
+		expect(transaction_data).to.have.own.property('maxPriorityFeePerGas');
 	});
 });
